+++ src/commands/mod.rs
@@ -1,32 +1,29 @@
 use crate::{Authentication, AuthenticationError};
 use prettytable::{cell, Cell, Row};
-use std::collections::HashMap;
 
 use log::debug;
-use std::fs;
-use std::fs::File;
-use std::io::Write;
-use std::path::Path;
-use std::str::FromStr;
 use std::time::Duration;
 
 use serde::{Deserialize, Deserializer, Serialize};
-use serde_json::json;
-use strum::IntoEnumIterator;
-use strum_macros::{EnumIter, EnumString, Display};
 
 use thiserror::Error;
 
 use reqwest::header::{HeaderMap, InvalidHeaderValue};
 use reqwest::StatusCode;
 
+pub use edge_app_settings::SettingType;
+pub use edge_app_settings::Setting;
+
 pub mod asset;
 pub mod edge_app;
+pub mod edge_app_manifest;
 pub(crate) mod edge_app_server;
+pub(crate) mod edge_app_settings;
 mod edge_app_utils;
 mod ignorer;
 pub(crate) mod playlist;
 pub mod screen;
+pub(crate) mod serde_utils;
 
 pub enum OutputType {
     HumanReadable,
@@ -221,185 +218,6 @@ pub fn patch<T: Serialize + ?Sized>(
     match serde_json::from_str(&response.text()?) {
         Ok(v) => Ok(v),
         Err(_) => Ok(serde_json::Value::Null),
-    }
-}
-
-fn string_field_is_none_or_empty(opt: &Option<String>) -> bool {
-    opt.as_ref().map_or(true, |s| s.is_empty())
-}
-
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
-pub struct EdgeAppManifest {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub app_id: Option<String>,
-    #[serde(
-        deserialize_with = "deserialize_option_string_not_empty",
-        skip_serializing_if = "string_field_is_none_or_empty",
-        default
-    )]
-    pub user_version: Option<String>,
-    #[serde(
-        deserialize_with = "deserialize_option_string_not_empty",
-        skip_serializing_if = "string_field_is_none_or_empty",
-        default
-    )]
-    pub description: Option<String>,
-    #[serde(
-        deserialize_with = "deserialize_option_string_not_empty",
-        skip_serializing_if = "string_field_is_none_or_empty",
-        default
-    )]
-    pub icon: Option<String>,
-    #[serde(
-        deserialize_with = "deserialize_option_string_not_empty",
-        skip_serializing_if = "string_field_is_none_or_empty",
-        default
-    )]
-    pub author: Option<String>,
-    #[serde(
-        deserialize_with = "deserialize_option_string_not_empty",
-        skip_serializing_if = "string_field_is_none_or_empty",
-        default
-    )]
-    pub homepage_url: Option<String>,
-    #[serde(
-        serialize_with = "serialize_settings",
-        deserialize_with = "deserialize_settings",
-        default
-    )]
-    pub settings: Vec<Setting>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Default, EnumString, Display, EnumIter)]
-pub enum SettingType {
-    #[default]
-    #[strum(serialize = "string")]
-    String,
-    #[strum(serialize = "secret")]
-    Secret,
-}
-
-// maybe we can use a better name as we have EdgeAppSettings which is the same but serde_json::Value inside
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
-pub struct Setting {
-    #[serde(
-        rename = "type", 
-        serialize_with = "serialize_setting_type",
-        deserialize_with = "deserialize_setting_type"
-    )]
-    pub type_: SettingType,
-    #[serde(default)]
-    pub default_value: String,
-    #[serde(default)]
-    pub title: String,
-    pub optional: bool,
-    pub help_text: String,
-}
-
-fn deserialize_settings<'de, D>(deserializer: D) -> Result<Vec<Setting>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let map: HashMap<String, Setting> = serde::Deserialize::deserialize(deserializer)?;
-    let mut settings: Vec<Setting> = map
-        .into_iter()
-        .map(|(title, mut setting)| {
-            setting.title = title;
-            setting
-        })
-        .collect();
-    settings.sort_by_key(|s| s.title.clone());
-    Ok(settings)
-}
-
-fn serialize_settings<S>(settings: &[Setting], serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    use serde::ser::SerializeMap;
-
-    let mut map = serializer.serialize_map(Some(settings.len()))?;
-    for setting in settings {
-        map.serialize_entry(&setting.title, &setting)?;
-    }
-    map.end()
-}
-
-fn deserialize_option_string_not_empty<'de, D>(deserializer: D) -> Result<Option<String>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let opt: Option<String> = Option::deserialize(deserializer)?;
-
-    match opt {
-        None => Ok(None),
-        Some(ref s) if s.is_empty() => Err(serde::de::Error::custom("String cannot be empty")),
-        _ => Ok(opt),
-    }
-}
-
-fn serialize_setting_type<S>(setting_type: &SettingType, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    serializer.serialize_str(&setting_type.to_string())
-}
-
-fn deserialize_setting_type<'de, D>(deserializer: D) -> Result<SettingType, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let s: String = Deserialize::deserialize(deserializer)?;
-
-    let valid_setting_types = SettingType::iter()
-        .map(|t| t.to_string())
-        .collect::<Vec<_>>()
-        .join("\n") + "\n";
-
-    match SettingType::from_str(&s.to_lowercase()) {
-        Ok(setting_type) => Ok(setting_type),
-        Err(_) => Err(serde::de::Error::custom(format!(
-            "Field type should be one of the following:\n{}",
-            valid_setting_types
-        ))),
-    }
-}
-
-impl EdgeAppManifest {
-    pub fn new(path: &Path) -> Result<Self, CommandError> {
-        let data = fs::read_to_string(path)?;
-        let manifest: EdgeAppManifest = serde_yaml::from_str(&data)?;
-        Ok(manifest)
-    }
-
-    pub fn save_to_file(manifest: &EdgeAppManifest, path: &Path) -> Result<(), CommandError> {
-        let yaml = serde_yaml::to_string(&manifest)?;
-        let manifest_file = File::create(path)?;
-        write!(&manifest_file, "---\n{yaml}")?;
-        Ok(())
-    }
-
-    fn prepare_payload(manifest: &EdgeAppManifest) -> HashMap<&str, serde_json::Value> {
-        [
-            ("app_id", &manifest.app_id),
-            ("user_version", &manifest.user_version),
-            ("description", &manifest.description),
-            ("icon", &manifest.icon),
-            ("author", &manifest.author),
-            ("homepage_url", &manifest.homepage_url),
-        ]
-        .iter()
-        .filter_map(|(key, value)| value.as_ref().map(|v| (*key, json!(v))))
-        .collect()
-    }
-
-    pub fn ensure_manifest_is_valid(path: &Path) -> Result<(), CommandError> {
-        match EdgeAppManifest::new(path) {
-            Ok(_) => Ok(()),
-            Err(e) => {
-                Err(CommandError::InvalidManifest(e.to_string()))
-            }
-        }
     }
 }
 
@@ -770,264 +588,6 @@ impl Formatter for PlaylistItems {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use tempfile::tempdir;
-
-    fn write_to_tempfile(dir: &tempfile::TempDir, file_name: &str, content: &str) -> std::path::PathBuf {
-        let file_path = dir.path().join(file_name);
-        std::fs::write(&file_path, content).unwrap();
-        file_path
-    }
-
-    #[test]
-    fn test_save_manifest_to_file_should_save_yaml_correctly() {
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("screenly.yml");
-
-        let manifest = EdgeAppManifest {
-            app_id: Some("test_app".to_string()),
-            user_version: Some("test_version".to_string()),
-            description: Some("test_description".to_string()),
-            icon: Some("test_icon".to_string()),
-            author: Some("test_author".to_string()),
-            homepage_url: Some("test_url".to_string()),
-            settings: vec![Setting {
-                title: "username".to_string(),
-                type_: SettingType::String,
-                default_value: "stranger".to_string(),
-                optional: true,
-                help_text: "An example of a setting that is used in index.html".to_string(),
-            }]
-        };
-
-        EdgeAppManifest::save_to_file(&manifest, &file_path).unwrap();
-
-        let contents = fs::read_to_string(file_path).unwrap();
-
-        let expected_contents = r#"---
-app_id: test_app
-user_version: test_version
-description: test_description
-icon: test_icon
-author: test_author
-homepage_url: test_url
-settings:
-  username:
-    type: string
-    default_value: stranger
-    title: username
-    optional: true
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        assert_eq!(contents, expected_contents);
-    }
-
-    #[test]
-    fn test_save_manifest_to_file_should_skip_none_optional_fields() {
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("screenly.yml");
-
-        let manifest = EdgeAppManifest {
-            app_id: Some("test_app".to_string()),
-            user_version: Some("test_version".to_string()),
-            description: None,
-            icon: Some("test_icon".to_string()),
-            author: None,
-            homepage_url: Some("test_url".to_string()),
-            settings: vec![Setting {
-                title: "username".to_string(),
-                type_: SettingType::String,
-                default_value: "stranger".to_string(),
-                optional: true,
-                help_text: "An example of a setting that is used in index.html".to_string(),
-            }]
-        };
-
-        EdgeAppManifest::save_to_file(&manifest, &file_path).unwrap();
-
-        let contents = fs::read_to_string(file_path).unwrap();
-
-        let expected_contents = r#"---
-app_id: test_app
-user_version: test_version
-icon: test_icon
-homepage_url: test_url
-settings:
-  username:
-    type: string
-    default_value: stranger
-    title: username
-    optional: true
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        assert_eq!(contents, expected_contents);
-    }
-
-    #[test]
-    fn test_save_manifest_to_file_should_skip_empty_optional_fields() {
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("screenly.yml");
-
-        let manifest = EdgeAppManifest {
-            app_id: Some("test_app".to_string()),
-            user_version: Some("test_version".to_string()),
-            description: Some("".to_string()),
-            icon: Some("test_icon".to_string()),
-            author: Some("".to_string()),
-            homepage_url: Some("test_url".to_string()),
-            settings: vec![Setting {
-                title: "username".to_string(),
-                type_: SettingType::String,
-                default_value: "stranger".to_string(),
-                optional: true,
-                help_text: "An example of a setting that is used in index.html".to_string(),
-            }]
-        };
-
-        EdgeAppManifest::save_to_file(&manifest, &file_path).unwrap();
-
-        let contents = fs::read_to_string(file_path).unwrap();
-
-        let expected_contents = r#"---
-app_id: test_app
-user_version: test_version
-icon: test_icon
-homepage_url: test_url
-settings:
-  username:
-    type: string
-    default_value: stranger
-    title: username
-    optional: true
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        assert_eq!(contents, expected_contents);
-    }
-
-    #[test]
-    fn test_save_manifest_to_file_should_skip_default_optional_fields() {
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("screenly.yml");
-
-        let manifest = EdgeAppManifest {
-            app_id: Some("test_app".to_string()),
-            settings: vec![Setting {
-                title: "username".to_string(),
-                type_: SettingType::String,
-                default_value: "stranger".to_string(),
-                optional: true,
-                help_text: "An example of a setting that is used in index.html".to_string(),
-            }],
-            ..Default::default()
-        };
-
-        EdgeAppManifest::save_to_file(&manifest, &file_path).unwrap();
-
-        let contents = fs::read_to_string(file_path).unwrap();
-
-        let expected_contents = r#"---
-app_id: test_app
-settings:
-  username:
-    type: string
-    default_value: stranger
-    title: username
-    optional: true
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        assert_eq!(contents, expected_contents);
-    }
-
-    #[test]
-    fn test_ensure_manifest_is_valid_when_file_non_existent_should_return_error() {
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("screenly.yml");
-    
-        assert!(EdgeAppManifest::ensure_manifest_is_valid(&file_path).is_err());
-    }
-
-    #[test]
-    fn test_ensure_manifest_is_valid_when_file_valid_should_return_ok() {
-        let dir = tempdir().unwrap();
-        let file_name = "screenly.yml";
-        let content = r#"---
-app_id: test_app
-settings:
-  username:
-    type: string
-    default_value: stranger
-    title: username
-    optional: true
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        write_to_tempfile(&dir, file_name, content);
-        let file_path = dir.path().join(file_name);        
-        assert!(EdgeAppManifest::ensure_manifest_is_valid(&file_path).is_ok());
-    }
-
-    #[test]
-    fn test_ensure_manifest_is_valid_when_missing_field_should_return_error() {
-        let dir = tempdir().unwrap();
-        let file_name = "screenly.yml";
-        let content = r#"---
-app_id: test_app
-settings:
-  username:
-    type: string
-    default_value: stranger
-    title: username
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        write_to_tempfile(&dir, file_name, content);
-        let file_path = dir.path().join(file_name);
-        assert!(EdgeAppManifest::ensure_manifest_is_valid(&file_path).is_err());
-    }
-
-    #[test]
-    fn test_ensure_manifest_is_valid_when_empty_field_should_return_error() {
-        let dir = tempdir().unwrap();
-        let file_name = "screenly.yml";
-        let content = r#"---
-app_id: test_app
-homepage_url: ''
-settings:
-  username:
-    type: string
-    default_value: stranger
-    title: username
-    optional: true
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        write_to_tempfile(&dir, file_name, content);
-        let file_path = dir.path().join(file_name);        
-        assert!(EdgeAppManifest::ensure_manifest_is_valid(&file_path).is_err());
-    }
-
-    #[test]
-    fn test_ensure_manifest_is_valid_when_invaild_type_should_return_error() {
-        let dir = tempdir().unwrap();
-        let file_name = "screenly.yml";
-        let content = r#"---
-app_id: test_app
-settings:
-  username:
-    type: bool
-    default_value: stranger
-    title: username
-    optional: true
-    help_text: An example of a setting that is used in index.html
-"#;
-
-        write_to_tempfile(&dir, file_name, content);
-        let file_path = dir.path().join(file_name);        
-        assert!(EdgeAppManifest::ensure_manifest_is_valid(&file_path).is_err());
-    }
 
     #[test]
     fn test_edge_app_versions_formatter_format_output_properly() {
@@ -1065,54 +625,6 @@ settings:
 +----------+-----------------+-----------+-------------------+
 "#
         );
-    }
-
-    #[test]
-    fn test_prepare_manifest_payload_includes_some_fields() {
-        let manifest = EdgeAppManifest {
-            app_id: Some("test_app".to_string()),
-            user_version: Some("test_version".to_string()),
-            description: Some("test_description".to_string()),
-            icon: Some("test_icon".to_string()),
-            author: Some("test_author".to_string()),
-            homepage_url: Some("test_url".to_string()),
-            settings: vec![Setting {
-                title: "username".to_string(),
-                type_: SettingType::String,
-                default_value: "stranger".to_string(),
-                optional: true,
-                help_text: "An example of a setting that is used in index.html".to_string(),
-            }]
-        };
-
-        let result = EdgeAppManifest::prepare_payload(&manifest);
-        assert_eq!(result["app_id"], json!("test_app"));
-        assert_eq!(result["user_version"], json!("test_version"));
-        assert_eq!(result["description"], json!("test_description"));
-        assert_eq!(result["icon"], json!("test_icon"));
-        assert_eq!(result["author"], json!("test_author"));
-        assert_eq!(result["homepage_url"], json!("test_url"));
-    }
-
-    #[test]
-    fn test_prepare_manifest_payload_omits_none_fields() {
-        let manifest = EdgeAppManifest {
-            app_id: Some("test_app".to_string()),
-            user_version: None,
-            description: Some("test_description".to_string()),
-            icon: Some("test_icon".to_string()),
-            author: None,
-            homepage_url: Some("test_url".to_string()),
-            ..Default::default()
-        };
-
-        let result = EdgeAppManifest::prepare_payload(&manifest);
-        assert_eq!(result["app_id"], json!("test_app"));
-        assert_eq!(result.contains_key("user_version"), false);
-        assert_eq!(result["description"], json!("test_description"));
-        assert_eq!(result["icon"], json!("test_icon"));
-        assert_eq!(result.contains_key("author"), false);
-        assert_eq!(result["homepage_url"], json!("test_url"));
     }
 }
